@@ -1,11 +1,20 @@
-import { ThemeProvider } from "@material-ui/core";
+import ThemeProvider from "@mui/material/styles/ThemeProvider";
 import { render, screen } from "@testing-library/react";
 import { useAppTheme } from "hooks/useAppTheme";
 import App from "./App";
 
-jest.mock("@material-ui/core");
-jest.mock("hooks/useAppTheme");
-jest.mock("components/MainLayout");
+jest.mock("@mui/material/styles/ThemeProvider", () => ({
+  __esModule: true,
+  default: jest.fn(),
+}));
+jest.mock("hooks/useAppTheme", () => ({
+  __esModule: true,
+  useAppTheme: jest.fn(),
+}));
+jest.mock("components/MainLayout", () => ({
+  __esModule: true,
+  default: jest.fn(),
+}));
 
 const mockedThemeProvider = ThemeProvider as jest.Mock;
 const mockedUseAppTheme = useAppTheme as jest.Mock;
@@ -19,6 +28,6 @@ test("renders ThemeProvider", () => {
   render(<App />);
   const linkElement = screen.getByText(/Weather app/i);
   expect(linkElement).toBeInTheDocument();
-  expect(mockedThemeProvider).toBeCalledTimes(1);
   expect(mockedUseAppTheme).toBeCalledTimes(1);
+  expect(mockedThemeProvider).toBeCalledTimes(1);
 });
