@@ -1,12 +1,14 @@
-import Container from "@material-ui/core/Container";
-import Grid from "@material-ui/core/Grid";
-import List from "@material-ui/core/List";
-import ListItem from "@material-ui/core/ListItem";
-import ListItemText from "@material-ui/core/ListItemText";
-import { makeStyles } from "@material-ui/core/styles";
+import Container from "@mui/material/Container";
+import Grid from "@mui/material/Grid";
+import List from "@mui/material/List";
+import ListItem from "@mui/material/ListItem";
+import ListItemText from "@mui/material/ListItemText";
+import Paper from "@mui/material/Paper";
+import { Theme } from "@mui/material/styles";
+import makeStyles from "@mui/styles/makeStyles";
 import { useFavorite } from "hooks/useFavorite";
 
-const useStyles = makeStyles((theme) => ({
+const useStyles = makeStyles<Theme>((theme) => ({
   container: {
     paddingTop: theme.spacing(4),
     paddingBottom: theme.spacing(4),
@@ -23,21 +25,25 @@ export default function Favorite() {
     <Container maxWidth="lg" className={classes.container}>
       <Grid container spacing={3}>
         <Grid item xs={12}>
-          <List component="nav" className={classes.list}>
-            {favoriteList.map((location) => (
-              <ListItem
-                key={location}
-                button
-                component="a"
-                href={`#/location/${location}`}
-              >
-                <ListItemText primary={location} />
-              </ListItem>
-            ))}
-            {favoriteList.length === 0 && (
-              <ListItem>No favorite, please add location in Home page</ListItem>
-            )}
-          </List>
+          <Paper elevation={2}>
+            <List component="nav" className={classes.list}>
+              {favoriteList.map((location) => (
+                <ListItem
+                  key={location}
+                  button
+                  component="a"
+                  href={`#/location/${location}`}
+                >
+                  <ListItemText primary={location} />
+                </ListItem>
+              ))}
+              {favoriteList.length === 0 && (
+                <ListItem>
+                  No favorite, please add location in Home page
+                </ListItem>
+              )}
+            </List>
+          </Paper>
         </Grid>
       </Grid>
     </Container>
