@@ -1,12 +1,13 @@
-import { makeStyles } from "@material-ui/core";
-import Container from "@material-ui/core/Container";
-import Grid from "@material-ui/core/Grid";
+import Container from "@mui/material/Container";
+import Grid from "@mui/material/Grid";
+import { Theme } from "@mui/material/styles";
+import makeStyles from "@mui/styles/makeStyles";
 import WeatherResult from "components/home/WeatherResult";
 import { useWeather } from "hooks/useWeather";
 import { useEffect, useState } from "react";
 import { useParams } from "react-router";
 
-const useStyles = makeStyles((theme) => ({
+const useStyles = makeStyles<Theme>((theme) => ({
   container: {
     paddingTop: theme.spacing(4),
     paddingBottom: theme.spacing(4),
@@ -19,8 +20,10 @@ export default function Location() {
   const { setLocation } = useWeather();
   const { location } = useParams<{ location: string }>();
   useEffect(() => {
-    setLocation(location);
-    setInit(true);
+    if (location != null) {
+      setLocation(location);
+      setInit(true);
+    }
   }, [setLocation, location]);
   return (
     <Container maxWidth="lg" className={classes.container}>
