@@ -6,7 +6,7 @@ module.exports = {
     "plugin:@typescript-eslint/eslint-recommended",
     "plugin:@typescript-eslint/recommended",
     "plugin:@typescript-eslint/recommended-requiring-type-checking",
-    "prettier",
+    "plugin:prettier/recommended",
   ],
   parserOptions: {
     tsconfigRootDir: __dirname,
