@@ -1,6 +1,6 @@
 import { bindActionCreators } from "@reduxjs/toolkit";
 import store from "store";
-import { updateSlice } from "store/slices/update";
+import { updateSlice } from "store/slices/update.slice";
 
 export const updateActions = bindActionCreators(
   updateSlice.actions,
