@@ -1,6 +1,6 @@
 import { bindActionCreators } from "@reduxjs/toolkit";
 import store from "store";
-import { weatherSlice } from "store/slices/weather";
+import { weatherSlice } from "store/slices/weather.slice";
 import * as weatherThunk from "store/thunks/weather";
 
 export const weatherActions = bindActionCreators(
