@@ -1,6 +1,6 @@
 import { useSelector } from "react-redux";
-import { State } from "store";
+import { RootState } from "store";
 
 export const useThemeType = () => {
-  return useSelector((state: State) => state.theme.themeType);
+  return useSelector((state: RootState) => state.theme);
 };
